Fixed issue with older StreamInfo objects.
Fixes #45

+++ src/info/ata4/unity/cli/extract/mesh/MeshData.java
@@ -267,7 +267,7 @@ class MeshData {
                     }
                 }
                 
-                in.align(stream.stride);
+                in.align(stream.stride.intValue());
             }
         }
     }
